huawei: add support for MU709
This gives us a functional MU709 stack.  The only thing missing is the
extended signal capabilities,  which are supported.

See HCSQ command in the AT command doc at:

    http://orange/hardware/ENG_DATA/Wireless_3G_LTE/Huawei/ME909u/

The HCSQ command and its unsolicitied responses can be used to provide
extended signal properties.  To add code to the Huawei plugin see
the ModemManager src/mm-broadband-modem-qmi.c file for an example.