port: prevent bearers from trying to use the same data port
When two bearers on a modem attempt to grab a data port to start their
connection setups, there is a chance they may grab the same data port.
As a result, ModemManager gets very confused and bad things happen.

Previously, the criteria for an available data port was only that it
was not yet connected. Now we require bearer connect and disconnect
procedures "claim" the data port during connect and disconnect
procedures.

The states during connection are:
	claimed -> connected -> unclaimed
The states during disconnection are:
	claimed -> disconnected -> unclaimed