qmi: nominate PDP profile indexes for bearer
Previously, ModemManager did not specify the "profile index" for each bearer
it tried to connect on the same modem. The profile index corresponds to the
entry in the AT+CGDCONT? list that the QMI commands would affect.

This meant that (in a dual APN setup) the bearers would all try to use the
"default profile number" (3 on Verizon firmware, 1 on others) which, in some
scenarios (locked to 4G?), would trigger disconnections.
E.g. Bearer 1 connects, Bearer 2 connects causing Bearer 1 to disconnect,
Bearer 1 reconnects causing Bearer 2 to disconnect, and so on.

Now, the user can nominate a profile index when creating the bearer e.g.
  mmcli ... --create-bearer='apn=vzwinternet,number=3'
and if not specified, the bearer finds the default profile number and uses
the chosen data port to offset it. E.g.:
  default profile number = 1
  wwan0 profile index = 1
  wwan1 profile index = 2