port-probe: avoid subsystem based checks when reading probe results
There should be no need to do an early check to filter out ports of
the wrong subsystem.

For the user of these methods it is irrelevant if the FALSE is
returned because the port is of the wrong subsystem, or because the
test wasn't added to be probed, or because the test actually failed.

In other words, the ports where the test succeeded will only have
succeeded if they are of the correct subsystem and if the test was
actually executed successfully.