altair: make sure parse_vendor_pco_info() always returns an error
The parse_vendor_pco_info() method was returning NULL without error if
the pco info string was empty.

Under this situation, the code would have tried to add a NULL MMPco
into the pco_list list, which is not desired.

Avoid this, by making sure a NULL return always sets an error.